Made a few minor changes to MailingList stuff
